Move `grain_compile` and `grain_dump` into `bin`.
